+++ server/src/database/mx_get_chats_by_name.c
@@ -1,6 +1,7 @@
 #include "database.h"
 
-int mx_get_chats_by_name(sqlite3_stmt *stmt, char *name, int max_rows, t_chat **chats ,int *result_count)
+int mx_get_chats_by_name(sqlite3_stmt *stmt, char *name, int max_rows, t_chat
+        **chats ,uint16_t *result_count)
 {
     int result = 0;
     MX_TRY_FUNCTION(sqlite3_reset(stmt), result);
